% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{AddOrReplaceNrrdHeaderField}
\alias{AddOrReplaceNrrdHeaderField}
\title{Add or replace lines in the header of a nrrd file}
\usage{
AddOrReplaceNrrdHeaderField(infile, outfile, newfields, Force = FALSE,
  Detached = FALSE, action = c("addreplace", "addonly", "replaceonly"))
}
\arguments{
\item{infile}{Path to input file}

\item{outfile}{Path to output file}

\item{newfields}{Named vector of fields to replace}

\item{Force}{Overwrite existing file (default FALSE)}

\item{Detached}{Write a detached header insted of a nrrd (default FALSE)}

\item{action}{addreplace (Default) addonly or replaceonly}
}
\value{
TRUE or FALSE depending on success
}
\description{
Input is a named vector of new fields. Use unnamed fields for comments.
Quote field names with spaces (conventionally with back ticks).
Note that this function will error out for invalid field names.
See http://teem.sourceforge.net/nrrd/format.html for nrrd field details
}
\examples{
\dontrun{
AddOrReplaceNrrdHeaderField(lhmaskfile,outfile=file.path(tmpdir,"LHMask.nrrd"),
  c("# My interesting comment",`space origin`="(2,2,2)"),Force=TRUE)
}
}
\seealso{
\code{\link{read.nrrd.header}}
}
