% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{Nrrd2op}
\alias{Nrrd2op}
\title{Binary operation on two nrrds, or on a nrrd and a constant}
\usage{
Nrrd2op(infiles, outfile, fun = c("max", "min", "+", "-", "x", "/"),
  gzip = FALSE, CreateDirs = TRUE, Verbose = TRUE, Force = FALSE)
}
\arguments{
\item{infiles}{Input files}

\item{outfile}{Output files}

\item{fun}{Binary function or operator}

\item{gzip}{Whether to gzip the resultant file}

\item{CreateDirs}{Whether to create implied output directories}

\item{Verbose}{Print status messages}

\item{Force}{Whether to overwrite an existing file}
}
\value{
The number of input files
}
\description{
Binary operation on two nrrds, or on a nrrd and a constant
}
