% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdCrc}
\alias{NrrdCrc}
\title{Find the CRC (hash) of a gzip encoded NRRD file}
\usage{
NrrdCrc(infile, UseGzip = FALSE, FastHeader = TRUE)
}
\arguments{
\item{infile}{Input/output files}

\item{UseGzip}{Whether to use the gzip command line tool to extract CRC
(default \code{FALSE}, see details)}

\item{FastHeader}{Parse only essential parts of NRRD header to save time.}
}
\description{
Find the CRC (hash) of a gzip encoded NRRD file
}
\details{
: Defaults to a quick method based on knowledge of gzip file format
from: \url{http://www.gzip.org/zlib/rfc-gzip.html} and assumption that there
is only one member in gzip data. Can also use gzip but this is much slower
since have to copy unu data to temporary file.
}
