% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdHisto}
\alias{NrrdHisto}
\title{Title}
\usage{
NrrdHisto(infile, outfile = sub("\\\\.([^.]+)$", ".histo.\\\\1", infile),
  maskfile, bins, min, max, blind8 = TRUE, ...)
}
\arguments{
\item{infile}{Input/output files}

\item{outfile}{Input/output files}

\item{maskfile}{Path to another image file defining which pixels in the input
file to include.}

\item{bins}{Number of bins for the histogram (see details)}

\item{min, max}{Minimum and maximum values for the histogram (see details)}

\item{blind8}{Whether to assume the full range (0,255) for 8 bit images
without checking}

\item{...}{Additional arguments passed to \code{\link{unu}} function.}
}
\value{
Output file
}
\description{
Make a histogram from a nrrd file (as a new nrrd)
}
\details{
\code{bins} defaults to 1000 for float/double image data, to the
  number of levels between \code{min} and \code{max} for integer data. If
  \code{min} or \code{max} is not specified, they will be set to the minimum
  (maximum) value found in the input data.
}
