% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{NrrdMakeDetachedHeaderForNrrd}
\alias{NrrdMakeDetachedHeaderForNrrd}
\title{Make a detached header for a specified nrrd file}
\usage{
NrrdMakeDetachedHeaderForNrrd(nrrd, nhdr = paste(nrrd, sep = ".",
  "nhdr"))
}
\arguments{
\item{nrrd}{Full path to a nrrd file}

\item{nhdr}{Full path nhdr file to be written}
}
\value{
invisibly returned character vector with new header
}
\description{
If nhdr is not supplied defaults to <nrrd>.nhdr.
If nhdr=NA then new header is returned but not written.
}
