% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdMerge}
\alias{NrrdMerge}
\title{Merge multiple NRRD images into a multichannel PNG}
\usage{
NrrdMerge(infiles, outdir = NULL, outfile = NULL, axis = 0,
  CreateDirs = TRUE, Verbose = TRUE, Force = FALSE,
  UseLock = FALSE, ...)
}
\arguments{
\item{infiles}{Input files}

\item{outdir}{Output directory}

\item{outfile}{Output files}

\item{axis}{0-indexed axis to join (merge) along (defaults to 0)}

\item{CreateDirs}{Whether to create implied output directories}

\item{Verbose}{Print status messages}

\item{Force}{Whether to overwrite an existing file}

\item{UseLock}{Whether to use file-based locking to enable simple
parallelisation}

\item{...}{Currently ignored}
}
\value{
Logical value indicating whether output exists
}
\description{
Merge multiple NRRD images into a multichannel PNG
}
