% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdMinMax}
\alias{NrrdMinMax}
\title{Return range of values in nrrd file (interface to unu minmax command)}
\usage{
NrrdMinMax(filename, blind8 = FALSE, ...)
}
\arguments{
\item{filename}{nrrd file containing data}

\item{blind8}{Assume 8 bit data has range 0-255 (or -127-127), default F.}

\item{...}{passed to \code{\link{system}} function}
}
\value{
c(min,max) or c(NA,NA) like R's \code{\link{range}} function
}
\description{
Return range of values in nrrd file (interface to unu minmax command)
}
\author{
jefferis
}
