% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdResample}
\alias{NrrdResample}
\title{Resample a nrrd with simple or complex filters}
\usage{
NrrdResample(infile, outfile, size, voxdims = NULL,
  centering = c("cell", "node"), otherargs = NULL, gzip = TRUE,
  suffix = NULL, CreateDirs = TRUE, Verbose = TRUE, Force = FALSE,
  UseLock = FALSE, ...)
}
\arguments{
\item{infile, outfile}{Input/output files}

\item{size}{Numeric vector of scale factors (NA=> don't touch) or integer
vector of pixel dimensions.}

\item{voxdims}{Target voxel dimensions}

\item{centering}{Whether to assume node or cell centering if not specified in
input file.}

\item{otherargs}{Passed to unu resample}

\item{gzip}{Whether or not to gzip encode the output file?}

\item{suffix}{A suffix added to the input file to construct an output
filename}

\item{CreateDirs}{Whether to create implied output directories}

\item{Verbose}{Print status messages}

\item{Force}{Whether to overwrite an existing file}

\item{UseLock}{Whether to use file-based locking to enable simple
parallelisation}

\item{...}{Additional params passed to unu}
}
\value{
The output file
}
\description{
wraps unu resample

  When size is explicitly an integer (e.g. c(512L,512L,100L) then it is
  assumed to represent the size of the desired output in pixels.

  unu resample defaults to cell centering in the asbence of information, but
  I prefer node.

  * downsampling a **cell** image (e.g. x2) will result in \itemize{

  \item an origin shift **iff** an origin was specified in the input file

  \item a change in the space directions that is not an exact multiple of the
  downsampling factor } Downsampling a node image (e.g. x2) will result in:
  \itemize{

  \item no change in origin

  \item a doubling in the space directions field }
}
\author{
jefferis
}
