% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdSave}
\alias{NrrdSave}
\title{Save a nrrd file in a different encoding / format}
\usage{
NrrdSave(infile, outfile, format = c("nrrd", "pnm", "text", "vtk", "png",
  "eps"), encoding = ifelse(format == "nrrd", "gzip", "raw"),
  CreateDirs = TRUE, UseLock = FALSE, DryRun = FALSE)
}
\arguments{
\item{infile, outfile}{Paths to input and output nrrds}

\item{format}{(one of 6 supported by unu save, default nrrd)}

\item{encoding}{(one of 5 supported by unu save, default gz when nrrd)}

\item{CreateDirs}{Whether to make missing direcories implied by output path}

\item{UseLock}{Whether to make a lockfile (useful for parallel processing)}

\item{DryRun}{Return command instead of running it}
}
\value{
path to ouput file
}
\description{
In outfile is missing, will overwrite infile
}
\seealso{
\code{\link{NrrdCrc}}
}
