% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdTestDataLength}
\alias{NrrdTestDataLength}
\title{Quick test that gzipped NRRD has correct data length}
\usage{
NrrdTestDataLength(infile, defaultReturnVal = TRUE)
}
\arguments{
\item{infile}{Input/output files}

\item{defaultReturnVal}{The value to return when file cannot be checked}
}
\value{
A logical indicating if the check passed
}
\description{
Quick test that gzipped NRRD has correct data length
}
\details{
Tests integrity of a nrrd file by checking that the data block is as
  long as it should be. For a gzip file, it checks that the last 4 bytes
  encode the length of the uncompressed data. This says nothing about the
  contents but does ensure that the file has not been truncated which is much
  the most common problem.
}
\seealso{
\code{\link{NrrdTestIntegrity}} which is more comprehensive but much
  slower
}
