% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdTestIntegrity}
\alias{NrrdTestIntegrity}
\title{Test integrity of a NRRD file using gzip/bzip2 checksum}
\usage{
NrrdTestIntegrity(infile, defaultReturnVal = TRUE)
}
\arguments{
\item{infile}{Input/output files}

\item{defaultReturnVal}{The value to return when file cannot be checked}
}
\value{
A logical indicating if the check passed
}
\description{
Test integrity of a NRRD file using gzip/bzip2 checksum
}
\seealso{
\code{\link{NrrdTestDataLength}} which is much faster
}
