% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtk.R
\name{resample_cmtk_dfield}
\alias{resample_cmtk_dfield}
\title{Resample a CMTK deformation field}
\usage{
resample_cmtk_dfield(infile, outfile = NULL, size = NULL,
  voxdims = NULL, ...)
}
\arguments{
\item{infile}{Input/output files}

\item{outfile}{Input/output files}

\item{size}{Numeric vector of scale factors (NA=> don't touch) or integer
vector of pixel dimensions.}

\item{voxdims}{Target voxel dimensions}

\item{...}{Additional params passed to unu}
}
\description{
Resample a CMTK deformation field
}
\examples{
\dontrun{
resample_cmtk_dfield("hemiBrain_JRC2018F_Warp_cmtk.nrrd", size=c(NA, rep(1/3, 3)))
# This currently fails because
resample_cmtk_dfield("hemiBrain_JRC2018F_Warp_cmtk.nrrd", voxdims=c(NA, 2,2,2))
}
}
