% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unu.R
\name{unu}
\alias{unu}
\alias{find_unu}
\title{Run the unu command line tool}
\usage{
unu(cmd = "", args = NULL, DryRun = FALSE, ...)

find_unu(location = getOption("runu.unu"))
}
\arguments{
\item{cmd}{The name of the unu command to run}

\item{args}{Additional arguments to unu}

\item{DryRun}{When \code{TRUE} show what would happen rather running command}

\item{...}{Additional arguments passed to \code{\link{system}}}

\item{location}{- location of the unu executable}
}
\value{
\code{unu} returns the value of \code{\link{system}} after running
  unu.
}
\description{
Run the unu command line tool
}
\details{
If \code{find_unu} cannot find your unu executable then either add
  the directory to your path or set the package option.
}
\section{Installation}{
 You will normally need to compile unu, but it is
  cross-platform (Windows/Mac/Linux) and has minimal dependencies. It is
  installed as part of the \bold{teem} libraries available at
  \url{http://teem.sourceforge.net/}. See in particular the
  \href{http://teem.sourceforge.net/download/index.html}{download} and
  \href{http://teem.sourceforge.net/build.html}{build} pages for details.
}

\examples{
\donttest{
find_unu()

# show help
unu()
}
\dontrun{
# manually set unu location if required
options(runu.unu="/path/to/unu")
}
}
