% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNUImageProcessingFunctions.R
\name{NrrdFlip}
\alias{NrrdFlip}
\title{Flip a NRRD image along one axis}
\usage{
NrrdFlip(infile, outfile, axes, suffix = NULL,
  endian = .Platform$endian, CreateDirs = TRUE, Verbose = TRUE,
  UseLock = FALSE, OverWrite = c("no", "update", "yes"))
}
\arguments{
\item{infile}{Input/output files}

\item{outfile}{Input/output files}

\item{axes}{Which axes to flip (0-indexed integer)}

\item{suffix}{Suffix to add to input file to construct output file.}

\item{endian}{Whether output image should be big or little endian for
multibyte data types (essentially all other than 8 bit). Defaults to
endianness of current machine.}

\item{CreateDirs}{Whether to create implied output directories}

\item{Verbose}{Print status messages}

\item{UseLock}{Whether to use file-based locking to enable simple
parallelisation}

\item{OverWrite}{Whether to overwrite (yes or no) or update output image.}
}
\value{
\code{TRUE} if output exists
}
\description{
Flip a NRRD image along one axis
}
